requirements.txt
pydantic
fastapi
sqlmodel
redis
fakeredis
quickjs
Unidecode
pykakasi
aiobotocore
Pillow
pydub
uvicorn